Create maze_runner.py
This program solves mazes using code from https://github.com/laurentluce/python-algorithms/tree/master/algorithms
Refer to http://www.laurentluce.com/posts/solving-mazes-using-python-simple-recursivity-and-a-search/

This puzzle involves the creation of a maze, this maze will be represented on the format: 10111;1110x 
In which: 1 = allowed path 0 = not allowed path/wall x = point that you need to find 
; = separation for next row of the maze 
The maze will be solved if you can find the shortest path that needs to be followed in order to get to the point! 
For example, for the maze: 
1100001;
0100111;
1111101;
1000111;
1110100;
1011110;
111001x
ALL the mazes will start from position 0,0.
You can only move to left, right, down and up, not in diagonal. 
Only the shortest path will be valid, on this case it will be: 
0.0, 0.1, 1.1, 2.1, 2.2, 2.3, 2.4, 3.4, 4.4, 5.4, 5.5, 6.5, 6.6